+++ coders/pnm.c
@@ -294,7 +294,7 @@ static Image *ReadPNMImage(const ImageInfo *image_info,ExceptionInfo *exception)
         image->storage_class=PseudoClass;
         image->colors=max_value+1;
       }
-    if ((size_t) (image->columns*image->rows) == 0)
+    if ((image->columns*image->rows) == 0)
       ThrowReaderException(CorruptImageWarning,
         "Unable to read image: image dimensions are zero",image);
     scale=(Quantum *) NULL;
