+++ coders/vicar.c
@@ -180,7 +180,7 @@ static Image *ReadVICARImage(const ImageInfo *image_info,
   length=0;
   image->columns=0;
   image->rows=0;
-  while (isgraph(c) && ((size_t) (image->columns*image->rows) == 0))
+  while (isgraph(c) && ((image->columns*image->rows) == 0))
   {
     if (!isalnum(c))
       {
